Consolidate E2E base classes into one base class
Since we're not going to be creating any other test fixtures derived
from SRTestBase, there's no point in having two base classes; it has
become over-engineering now that we've decided on just one E2E test.
